update session state machines
