+++ Sources/WalletConnect/Engine/SessionStateMachines/NonControllerSessionStateMachine.swift
@@ -2,13 +2,13 @@
 import Foundation
 import WalletConnectUtils
 import WalletConnectKMS
+import Combine
 
 final class NonControllerSessionStateMachine: SessionStateMachineValidating {
     
     var onMethodsUpdate: ((String, Set<String>)->())?
     
     private let sequencesStore: SessionSequenceStorage
-    private let wcSubscriber: WCSubscribing
     private let relayer: WalletConnectRelaying
     private let kms: KeyManagementServiceProtocol
     private var publishers = [AnyCancellable]()
@@ -16,28 +16,24 @@ final class NonControllerSessionStateMachine: SessionStateMachineValidating {
 
     init(relay: WalletConnectRelaying,
          kms: KeyManagementServiceProtocol,
-         subscriber: WCSubscribing,
          sequencesStore: SessionSequenceStorage,
-         metadata: AppMetadata,
-         logger: ConsoleLogging,
-         topicGenerator: @escaping () -> String = String.generateTopic) {
+         logger: ConsoleLogging) {
         self.relayer = relay
         self.kms = kms
-        self.wcSubscriber = subscriber
         self.sequencesStore = sequencesStore
         self.logger = logger
         setUpWCRequestHandling()
     }
     
     private func setUpWCRequestHandling() {
-        wcSubscriber.onReceivePayload = { [unowned self] subscriptionPayload in
+        relayer.wcRequestPublisher.sink { [unowned self] subscriptionPayload in
             switch subscriptionPayload.wcRequest.params {
             case .sessionUpdateMethods(let updateParams):
                 wcSessionUpdateMethods(payload: subscriptionPayload, updateParams: updateParams)
             default:
                 logger.warn("Warning: Session Engine - Unexpected method type: \(subscriptionPayload.wcRequest.method) received from subscriber")
             }
-        }
+        }.store(in: &publishers)
     }
     
     private func wcSessionUpdateMethods(payload: WCRequestSubscriptionPayload, updateParams: SessionType.UpdateMethodsParams) {
@@ -57,19 +53,5 @@ final class NonControllerSessionStateMachine: SessionStateMachineValidating {
         sequencesStore.setSequence(session)
         relayer.respondSuccess(for: payload)
         onMethodsUpdate?(session.topic, updateParams.methods)
-    }
-}
-
-protocol SessionStateMachineValidating {
-    func validateMethods(_ methods: Set<String>) -> Bool
-}
-
-extension SessionStateMachineValidating {
-    func validateMethods(_ methods: Set<String>) -> Bool {
-        for method in methods {
-            if method.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty {
-                return false
-            }
-        }
     }
 }
